api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.25.2
bcrypt==4.1.2
python-dotenv==1.0.0 